Revert "Merge pull request #13 from aws-neuron/release_2.21"
This reverts commit a694ac05903c53d311a299bfaa59c812618f57d2.
