+++ docs/general/config_overview.rst
@@ -59,7 +59,6 @@ and this key allows users to configure the ``trainer``.
     limit_val_batches: 1
     limit_test_batches: 1
     gradient_clip_val: 1.0
-    sequential_move_factor: 11
 
 .. note::
 
@@ -135,16 +134,6 @@ Float value to clip gradients at.
 
     * **Type**: float
     * **Required**: True
-
-
-**sequential_move_factor**
-
-Number of ranks/devices participating in initializing the model weights in parallel. Useful to reduce init time
-when using TP-PP config. The value can be increased upto the number of ``trainer.devices`` being used.
-
-    * **Default value**: 11
-    * **Type**: integer
-    * **Required**: False
 
 
 
@@ -487,23 +476,6 @@ This needs to be set if users want to use the
 
     * **Type**: bool
     * **Required**: True
-
-**fuse_qkv**
-
-This is set if users want to use fused q, k and v tensors in
-`GQAQKVLinear module <https://awsdocs-neuron.readthedocs-hosted.com/en/latest/libraries/neuronx-distributed/api-reference-guide-training.html#gqa-qkv-linear-module>`_ Using fuse_qkv can improve throughput.
-This parameter is True by default.
-
-    * **Type**: bool
-    * **Required**: False
-
-**pipeline_cuts**
-
-This is set as a list of layer names if users want to specify manual cut points for pipeline parallelism.
-One example is ['model.layers.10', 'model.layers.20'] in the case of PP=3.
-
-    * **Type**: List[str]
-    * **Required**: False
 
 **use_cpu_initialization**
