+++ docs/general/installation_guide.rst
@@ -32,7 +32,6 @@ Install the neuron packages using the command:
 
 .. code-block :: shell
 
-    pip install -U pip
     pip install --upgrade neuronx-cc==2.* torch-neuronx torchvision neuronx_distributed --extra-index-url https://pip.repos.neuron.amazonaws.com
 
 Building Apex
@@ -100,9 +99,6 @@ We can download the ``requirements.txt`` using the command:
 .. code-block :: shell
 
     wget https://raw.githubusercontent.com/aws-neuron/neuronx-distributed-training/master/requirements.txt
-
-If you are using a U22 or later OS, please see `Common failures during installation`_,
-as these errors can happen during requirements installation.
 
 We can now install the dependencies of the library using the following command:
