+++ docs/general/known_issues.rst
@@ -160,13 +160,3 @@ Flash Attention not supported for megatron-style models
 
 Flash attention kernel is supported only for HF-style models and will be added for megatron-style models in one of
 the future releases.
-
-During Dataset Download
-#######################
-
-If you are seeing an issue when downloading the datasets from the tutorials such as
-
-::
-    The maximum number of fds has been exceeded.
-
-We just need to increase the number of file descriptors with: `ulimit -n 65535`.
