+++ docs/tutorials/index.rst
@@ -9,7 +9,5 @@ This section will go over a bunch of tutorials to help users get started with Ne
     :maxdepth: 1
 
     Megatron GPT Pretraining <megatron_gpt_pretraining>
-    HuggingFace Llama3-8B Pretraining <hf_llama3_8B_pretraining>
-    HuggingFace Llama3-8B Supervised Fine-tuning <hf_llama3_8B_SFT>
-    HuggingFace Llama3-8B Direct Preference Optimization (DPO) based Fine-tuning <hf_llama3_8B_DPO>
+    HuggingFace LLama3-8B Supervised Fine-tuning <hf_llama3_8B_SFT>
     Checkpoint Conversion <checkpoint_conversion>
