+++ examples/conf/hf_llama3_70B_config.yaml
@@ -72,7 +72,7 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: selective # 'selective' or 'full'
-  activations_checkpoint_recompute: [CoreAttention, LlamaMLP]
+  activations_checkpoint_recompute_mlp: True # will be restructured in future
 
   fusions:
     softmax: True
