+++ examples/conf/hf_llama3_8B_SFT_config.yaml
@@ -82,7 +82,6 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: selective # 'selective' or 'full'
-  activations_checkpoint_recompute: [CoreAttention]
 
   fusions:
     softmax: True
