+++ examples/conf/hf_llama3_8B_config.yaml
@@ -72,7 +72,6 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: selective # 'selective' or 'full'
-  activations_checkpoint_recompute: [CoreAttention]
 
   fusions:
     softmax: True
