+++ examples/conf/megatron_llama2_7B_SFT_config.yaml
@@ -112,7 +112,6 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: selective # 'selective' or 'full'
-  activations_checkpoint_recompute: [CoreAttention]
 
   fusions:
     softmax: True
