+++ examples/conf/megatron_llama_70b_config.yaml
@@ -120,7 +120,6 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: full # 'selective' or 'full'
-  activations_checkpoint_recompute: null
 
   fusions:
     softmax: True
