+++ examples/conf/megatron_mistral_config.yaml
@@ -121,7 +121,6 @@ model:
 
   ## Activation Checkpointing
   activations_checkpoint_granularity: selective # 'selective' or 'full'
-  activations_checkpoint_recompute: [CoreAttention]
 
   fusions:
     softmax: True
