--- requirements_dpo.txt
@@ -1,43 +1,0 @@
-hydra-core>=1.2.0,<1.3
-omegaconf>=2.2,<2.3
-pyyaml==6.0.1
-torchmetrics>=0.4.1rc0,<=0.10.3
-trl==0.10.1
-transformers==4.31.0
-wandb
-webdataset>=0.1.48,<=0.1.62
-pandas
-sentencepiece
-youtokentome
-h5py
-ijson
-matplotlib>=3.3.2
-sacremoses
-sacrebleu
-einops
-faiss-cpu
-sentence_transformers<=2.2.2
-nltk>=3.6.5
-jieba
-ftfy
-gdown
-inflect
-jieba
-opencc==1.1.6
-pangu
-rapidfuzz
-pybind11
-pytorch-lightning==1.8.6
-ipadic
-mecab-python3
-protobuf==3.20.3
-datasets
-dill==0.3.8
-nemo_toolkit==1.14
-regex
-requests<2.32.0
-python-daemon
-huggingface_hub>=0.14.1,<0.24.0
-multiprocess==0.70.16
-numba<=0.60.0
-numpy>=1.24.3,<=1.25.2
