+++ src/neuronx_distributed_training/lightning_modules/model/hf_models/base_model.py
@@ -13,19 +13,6 @@ from ..base import BaseModelModule
 
 
 class BaseHfModel(BaseModelModule):
-    def on_train_start(self):
-        if 'dpo' in getattr(self.config.data, 'alignment_strategy', {}):
-            self.dpo_module = DPOBaseModel(self.config, self.trainer, self.model)
-            self.dpo_module.on_train_start(self.trainer, self.model, self.config)
-        else:
-            super().on_train_start()
-
-    def model_fwd_calc_loss(self, batch):
-        if 'dpo' in getattr(self.config.data, 'alignment_strategy', {}):
-            return self.dpo_module.model_fwd_calc_loss(self.model, batch, self.config)
-        else:
-            return super().model_fwd_calc_loss(batch)
-    
     def build_model(self):
         return self._get_model()
