+++ src/neuronx_distributed_training/lightning_modules/model/hf_models/llama_model.py
@@ -4,17 +4,13 @@
 import neuronx_distributed as nxd
 import torch
 from transformers import LlamaConfig
-from transformers.models.llama.modeling_llama import LlamaRotaryEmbedding
-import sys
-from neuronx_distributed.utils.utils import hardware
-from torch_neuronx.utils import get_platform_target
+
 from neuronx_distributed_training.models.hf_models.modeling_llama import (
     CoreAttention,
     LlamaDecoderLayer,
     LlamaForCausalLM,
     LlamaRMSNorm,
-    LlamaMLP,
-    ActivationMultiplyMLP
+    LlamaMLP
 )
 
 from .base_model import BaseHfModel
@@ -30,11 +26,6 @@ class HFLLamaModule(BaseHfModel):
         config.kv_shared_group_size = self.config.distributed_strategy.get("kv_replicator", 1)
         config.max_position_embeddings = max(config.max_position_embeddings, self.config.model.get("max_position_embeddings"))
         config.use_flash_attention = self.config.model.fusions.flash_attention
-        hardware_type = hardware(get_platform_target())
-        if hardware_type==hardware.TRN1:
-            config.lnc = self.config.trainer.get("lnc", 1)
-        if hardware_type==hardware.TRN2:
-            config.lnc = self.config.trainer.get("lnc", 2)
         if self.config.model.get('num_layers', -1) != -1:
             config.num_hidden_layers = self.config.model.get('num_layers')
         if self.config.model.get('hidden_size', -1) != -1:
@@ -43,23 +34,13 @@ class HFLLamaModule(BaseHfModel):
             config.rope_theta = self.config.model.get('rope_theta')
 
         leaf_module_cls = [LlamaRMSNorm.__name__]
-        activation_recompute_modules = []
-        recompute_modules = self.config.model.get("activations_checkpoint_recompute", [])
-        granularity = self.config.model.get("activations_checkpoint_granularity", None)
-
-        if granularity == "selective":
-            for module in recompute_modules:
-                module_obj = getattr(sys.modules[__name__], module, None)
-                if module_obj is not None:
-                    activation_recompute_modules.append(module_obj)
-        elif granularity == "full":
-            activation_recompute_modules = "full"
-        elif not self.config.model.fusions.get("flash_attention", False):
-            activation_recompute_modules.append(CoreAttention) # do CoreAttention checkpointing if flash_attention is off
-        else:
-            activation_recompute_modules = None
-
-        self.nxd_config["activation_checkpoint_config"] = activation_recompute_modules
+        if self.config.model.get("activations_checkpoint_granularity", None) == "selective":
+            if self.config.model.get("activations_checkpoint_recompute_mlp", False) and self.config.model.encoder_seq_length>=8192:
+                self.nxd_config["activation_checkpoint_config"] = (CoreAttention, LlamaMLP)
+            else:
+                self.nxd_config["activation_checkpoint_config"] = CoreAttention
+        elif self.config.model.get("activations_checkpoint_granularity", None) == "full":
+            self.nxd_config["activation_checkpoint_config"] = "full"
         self.nxd_config["pipeline_config"].update(
             {
                 "transformer_layer_cls": LlamaDecoderLayer,
