+++ src/neuronx_distributed_training/lightning_modules/nlp_overrides.py
@@ -104,7 +104,6 @@ from torch import Tensor
 from torch.utils.data import DataLoader
 from torch_xla.distributed.zero_redundancy_optimizer import ZeroRedundancyOptimizer
 from torchmetrics import Metric
-from neuronx_distributed_training.utils import get_lnc_size
 
 
 def has_len_all_ranks_patched(
@@ -658,14 +657,10 @@ class NLPTrainer(Trainer):
         move_metrics_to_cpu: bool = False,
         multiple_trainloader_mode: str = "max_size_cycle",
         inference_mode: bool = True,
-        lnc: int = None,
-        sequential_move_factor: Optional[int] = None,
     ) -> None:
         Trainer._log_api_event("init")
         logging.info(f"{self.__class__.__name__}: Initializing trainer with parameters: {locals()}")
         self.state = TrainerState()
-        self.lnc = get_lnc_size(lnc)
-        self.sequential_move_factor = sequential_move_factor
 
         if default_root_dir is not None:
             default_root_dir = os.fspath(default_root_dir)
