+++ src/neuronx_distributed_training/utils/__init__.py
@@ -1,4 +1,4 @@
 # Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from .utils import Throughput, get_lnc_size
+from .utils import Throughput
