+++ src/neuronx_distributed_training/utils/utils.py
@@ -3,17 +3,7 @@
 
 import queue
 import time
-from neuronx_distributed.utils.utils import hardware
-from torch_neuronx.utils import get_platform_target
 
-def get_lnc_size(lnc):
-    hardware_type = hardware(get_platform_target())
-    if hardware_type == hardware.TRN2:
-        if lnc is None:
-            lnc = 2
-    else:
-        lnc = 1
-    return lnc
 
 class Throughput:
     def __init__(self, moving_avg_window_size):
